Rename the package to Slender.Scanner
#9
