+++ README.md
@@ -1,2 +1,2 @@
-# AssemblyScanner
+# Scanner
 A simple .NET assembly scanner and associated visitor
