--- sample/Slender.AssemblyScanner.Sample/Shapes/Circle.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Shapes
-{
-
-    public class Circle : IShape { }
-
-}
