--- sample/Slender.AssemblyScanner.Sample/Shapes/IShape.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Shapes
-{
-
-    public interface IShape { }
-
-}
