--- sample/Slender.AssemblyScanner.Sample/Shapes/Rectangle.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Shapes
-{
-
-    public class Rectangle : IShape { }
-
-}
