--- sample/Slender.AssemblyScanner.Sample/Shapes/Triangle.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Shapes
-{
-
-    public class Triangle : IShape { }
-
-}
