--- sample/Slender.AssemblyScanner.Sample/Vehicles/Boat.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Vehicles
-{
-
-    public class Boat : Watercraft { }
-
-}
