--- sample/Slender.AssemblyScanner.Sample/Vehicles/Car.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Vehicles
-{
-
-    public class Car : RoadVehicle { }
-
-}
