--- sample/Slender.AssemblyScanner.Sample/Vehicles/JetSki.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Vehicles
-{
-
-    public class JetSki : Watercraft { }
-
-}
