--- sample/Slender.AssemblyScanner.Sample/Vehicles/Motorbike.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Vehicles
-{
-
-    public class Motorbike : RoadVehicle { }
-
-}
