--- sample/Slender.AssemblyScanner.Sample/Vehicles/Truck.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Vehicles
-{
-
-    public class Truck : RoadVehicle { }
-
-}
