--- sample/Slender.AssemblyScanner.Sample/Vehicles/Vehicle.cs
@@ -1,6 +1,0 @@
-﻿namespace Slender.AssemblyScanner.Sample.Vehicles
-{
-
-    public abstract class Vehicle { }
-
-}
