+++ sample/Slender.Scanner.Sample/Program.cs
@@ -1,5 +1,5 @@
-﻿using Slender.AssemblyScanner;
-using Slender.AssemblyScanner.Sample;
+﻿using Slender.Scanner;
+using Slender.Scanner.Sample;
 
 var _AssemblyScan = AssemblyScan.FromAssembly(typeof(Program).Assembly);
 var _Visitor = new LoggingVisitor();
