+++ sample/Slender.Scanner.Sample/Shapes/Circle.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Shapes
+{
+
+    public class Circle : IShape { }
+
+}
