+++ sample/Slender.Scanner.Sample/Shapes/IShape.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Shapes
+{
+
+    public interface IShape { }
+
+}
