+++ sample/Slender.Scanner.Sample/Shapes/Rectangle.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Shapes
+{
+
+    public class Rectangle : IShape { }
+
+}
