+++ sample/Slender.Scanner.Sample/Shapes/Triangle.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Shapes
+{
+
+    public class Triangle : IShape { }
+
+}
