+++ sample/Slender.Scanner.Sample/Slender.Scanner.Sample.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Slender.AssemblyScanner\Slender.AssemblyScanner.csproj" />
+    <ProjectReference Include="..\..\src\Slender.Scanner\Slender.Scanner.csproj" />
   </ItemGroup>
 
 </Project>
