+++ sample/Slender.Scanner.Sample/Vehicles/Boat.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Vehicles
+{
+
+    public class Boat : Watercraft { }
+
+}
