+++ sample/Slender.Scanner.Sample/Vehicles/Car.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Vehicles
+{
+
+    public class Car : RoadVehicle { }
+
+}
