+++ sample/Slender.Scanner.Sample/Vehicles/JetSki.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Vehicles
+{
+
+    public class JetSki : Watercraft { }
+
+}
