+++ sample/Slender.Scanner.Sample/Vehicles/Motorbike.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Vehicles
+{
+
+    public class Motorbike : RoadVehicle { }
+
+}
