+++ sample/Slender.Scanner.Sample/Vehicles/Truck.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Vehicles
+{
+
+    public class Truck : RoadVehicle { }
+
+}
