+++ sample/Slender.Scanner.Sample/Vehicles/Vehicle.cs
@@ -1,0 +1,6 @@
+﻿namespace Slender.Scanner.Sample.Vehicles
+{
+
+    public abstract class Vehicle { }
+
+}
