+++ src/Slender.Scanner/AssemblyScan.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace Slender.AssemblyScanner
+namespace Slender.Scanner
 {
 
     /// <summary>
