+++ src/Slender.Scanner/AssemblyScanVisitor.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Slender.AssemblyScanner
+namespace Slender.Scanner
 {
 
     /// <summary>
