+++ src/Slender.Scanner/Slender.Scanner.csproj
