+++ test/Slender.Scanner.Tests/Unit/AssemblyScanVisitorTests.cs
@@ -1,12 +1,12 @@
 ﻿using FluentAssertions;
 using Moq;
-using Slender.AssemblyScanner.Tests.Support;
+using Slender.Scanner.Tests.Support;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 
-namespace Slender.AssemblyScanner.Tests.Unit
+namespace Slender.Scanner.Tests.Unit
 {
 
     public class AssemblyScanVisitorTests
